requirements.txt
# Librerías principales para análisis de señales ECG
wfdb==4.1.0
numpy>=1.24.3
pandas>=1.0.0,<2.0.0
scipy>=1.11.1

# Visualización
matplotlib>=3.7.2
seaborn>=0.12.2
plotly>=5.15.0

# Machine Learning tradicional
scikit-learn>=1.3.0

# Deep Learning - Optimizado para Apple Silicon M1
# Para M1 Mac: usar tensorflow-macos y tensorflow-metal
tensorflow-macos>=2.13.0
tensorflow-metal>=1.0.0
keras>=2.13.1

# Procesamiento de señales
PyWavelets>=1.4.0
# scikit-sparse>=0.4.8  # Opcional - puede tener problemas de compilación en algunos sistemas

# Análisis de señales biomédicas
heartpy>=1.2.7
nolds>=0.5.2

# Utilidades adicionales
tqdm>=4.65.0
requests>=2.31.0
joblib>=1.3.0

# Jupyter y desarrollo
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=6.5.4

# Caché y optimización
cachetools>=5.3.0
